requirements.txt
numpy==1.23.5
flask
flask_cors
transformers
tensorflow==2.12.0
sentencepiece
requests
flask_sqlalchemy
psycopg2-binary
python-dotenv
gdown
setuptools
wheel
flask-bcrypt
python-dotenv
PyJWT
pymongo
pytz
Werkzeug
langdetect
googletrans==4.0.0rc1
